requirements.txt
python-telegram-bot
lottie==0.6.0